Combat update (example in testbed scene)
Combat state machine begun. Combat parts in Player.cs restructured a bit
to work with it.

Characters pause combat for a second upon being hit.

Enemies can be killed and removed as targets.

Target reticle resets if it's on a dead enemy.

Combat can now be "won" when enemies are defeated, including returning
to exploration mode.