Animation node, load start node, Winter crab functionality
CM_StartOnLoad complete. One second after the scene loads, this triggers
and activates down the chain.

CM_PlayAnimation and CM_StopAnimation created, allowing cutscenes to
manually play animations. Note that this forces the character object to
completely ignore everything else and play an animation, meaning you
have to call CM_StopAnimation to go back to normal behavior.

Winter crab's animation is in, and is ready to be plugged in.