Cactus broke, temp fix, Knossis backgrounds and wight sheets imported
Second cactus image broke, despite making no sense to anyone (half of
the sprite is sliced off when in a scene for no discernable reason
whatsoever, in and outside of this project). Replaced with first cactus,
flipped for now.

Imported new Knossis backgrounds.

Wraith Brawler, Fencer, Gunner imported (but not yet implimented.)