SceneFlagManager update
SceneFlagManager added. When a battleground is wiped by the player, the
SceneFlagManager is updated, and sets an appropriate local flag. When
the scene is later loaded by the player, the manager looks at each
battleground in a level and compares it to these local flags; if a
battleground has been wiped previously, it disables that battleground.
These local variables are to be reset when a new area is loaded - the
functionality is present, but is not implimented quite yet.

SceneFlagManager looks at the AreaInfo flag when it loads; if the entry
is blank, it resumes normally. If it has an entry, it automatically
updates the camera to use it. (This AreaInfo flag needs to be set when
the player saves.)